Lekhan_Gowda_R_24008_023_Steganography.zip
Image Steganography using LSB Encoding and Decoding:  The objective was to develop a method to securely transmit a secret text file by encoding it within an image of BMP file format. The encoding process involved storing the length of the hidden text in the image, followed by embedding the text data bit by bit into the LSB of the image bytes. The decoding process retrieves the encoded length and then extracts the hidden text data from the image, producing the original secret message after decoding.