leku car black box.X.zip
Car Black Box using PIC Microcontroller for Event Logging in EEPROM: The objective was to create a basic black box system to log critical events, such as over-speeding, for vehicle post-crash analysis. Events were continuously monitored, and data was logged in EEPROM. The system can be adapted for use in various types of vehicles.